refactor(bootstrap): simplify imports and improve logger handling
